Update and rename iAmTryingMyHardestToUnderstandThisGoddamnLanguageButIcan'tSoIamSusmittingSomethingSoThatYouCanSeeThatI'mTrying.js to SumOnJS.js
